# api/panel-session.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HomeLab Harness Panel session",
    "version": "1.1.0",
    "description": "Single-owner browser session in front of the Harness API. The public edge authenticates the operator and overwrites X-Panel-Authenticated-User. Panel derives the authoritative ownerId from the signed Harness registry; it never accepts ownerId or a YouTrack token from the browser."
  },
  "paths": {
    "/api/v2/healthz": {
      "get": {
        "summary": "Process liveness",
        "responses": {
          "200": {
            "description": "Panel is running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["panel"],
                  "properties": { "panel": { "const": "up" } }
                }
              }
            }
          }
        }
      },
      "ConfigurationDiagnostic": { "type": "object", "additionalProperties": false, "required": ["code", "pointer", "line", "column", "message"], "properties": { "code": { "type": "string" }, "pointer": { "type": "string" }, "line": { "type": "integer", "minimum": 1 }, "column": { "type": "integer", "minimum": 1 }, "message": { "type": "string" } } },
      "ConfigurationValidationState": { "type": "object", "additionalProperties": false, "required": ["valid", "diagnostics", "effectStatus"], "properties": { "valid": { "type": "boolean" }, "diagnostics": { "type": "array", "items": { "$ref": "#/components/schemas/ConfigurationDiagnostic" } }, "effectStatus": { "const": "none" } } },
      "BuildContextAssetRef": { "type": "object", "additionalProperties": false, "required": ["path", "assetId", "sha256"], "properties": { "path": { "type": "string", "minLength": 1, "maxLength": 256 }, "assetId": { "type": "string", "format": "uuid" }, "sha256": { "type": "string", "pattern": "^[0-9a-f]{64}$" } } },
      "BuildContextManifest": { "type": "object", "additionalProperties": false, "required": ["revision", "assets"], "properties": { "revision": { "type": "string", "format": "uuid" }, "assets": { "type": "array", "maxItems": 1000, "items": { "$ref": "#/components/schemas/BuildContextAssetRef" } } } },
      "ConfigurationValidateInput": { "type": "object", "additionalProperties": false, "required": ["schemaId", "rawJsonText", "rawDockerfileText"], "properties": { "schemaId": { "const": "agent-configuration-validate-v2" }, "rawJsonText": { "type": "string", "description": "UTF-8, at most 256 KiB decoded" }, "rawDockerfileText": { "type": "string", "description": "UTF-8, at most 128 KiB decoded" }, "buildContextManifest": { "$ref": "#/components/schemas/BuildContextManifest" } } },
      "ConfigurationSaveInput": { "type": "object", "additionalProperties": false, "required": ["schemaId", "expectedDraftVersion", "rawJsonText", "rawDockerfileText"], "properties": { "schemaId": { "const": "agent-configuration-save-v2" }, "expectedDraftVersion": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 }, "rawJsonText": { "type": "string" }, "rawDockerfileText": { "type": "string" }, "buildContextManifest": { "$ref": "#/components/schemas/BuildContextManifest" } } },
      "ConfigurationValidationResponse": { "type": "object", "additionalProperties": false, "required": ["schemaId", "validation"], "properties": { "schemaId": { "const": "agent-configuration-validate-v2" }, "buildContextManifest": { "$ref": "#/components/schemas/BuildContextManifest" }, "validation": { "$ref": "#/components/schemas/ConfigurationValidationState" } } },
      "ConfigurationDraft": { "type": "object", "additionalProperties": false, "required": ["schemaId", "nodeId", "draftVersion", "rawJsonText", "rawDockerfileText", "validation"], "properties": { "schemaId": { "const": "agent-configuration-draft-v2" }, "nodeId": { "type": "string", "format": "uuid" }, "draftVersion": { "type": "integer", "minimum": 1 }, "rawJsonText": { "type": "string" }, "rawDockerfileText": { "type": "string" }, "buildContextManifest": { "$ref": "#/components/schemas/BuildContextManifest" }, "validation": { "$ref": "#/components/schemas/ConfigurationValidationState" } } }
    },
    "/api/v2/bootstrap": {
      "post": {
        "summary": "Exchange trusted edge authentication for an opaque Panel session",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "format": "uri" }
          },
          {
            "name": "X-Panel-Authenticated-User",
            "in": "header",
            "required": true,
            "description": "Injected and overwritten by the authenticated reverse proxy; never supplied by browser code.",
            "schema": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._@-]{0,127}$" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object", "additionalProperties": false } }
          }
        },
        "responses": {
          "200": {
            "description": "Session issued",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Session" } }
            }
          },
          "403": { "description": "Origin or trusted edge authentication missing" },
          "429": { "description": "Bounded bootstrap capacity exhausted" }
        }
      }
    },
    "/api/v2/session": {
      "get": {
        "summary": "Recover the current Panel session and CSRF nonce",
        "security": [{ "panelSession": [] }],
        "responses": {
          "200": {
            "description": "Current session",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Session" } }
            }
          },
          "401": { "description": "Session missing or expired" }
        }
      }
    },
    "/api/v2/agents": {
      "get": {
        "summary": "Read the authenticated owner's agent-service inventory",
        "security": [{ "panelSession": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string", "minLength": 1, "maxLength": 512 }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner-scoped inventory page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./agent-service-v1.openapi.json#/components/schemas/InventoryPage"
                }
              }
            }
          },
          "400": { "description": "Invalid query or cursor" },
          "401": { "description": "Session missing or expired" },
          "404": { "description": "agent-service is not configured" },
          "503": { "description": "agent-service or its database is unavailable" }
        }
      }
    },
    "/api/v2/agents/{nodeId}/dialogs": {
      "get": {
        "summary": "Read permanent dialog bindings for one exact agent",
        "security": [{ "panelSession": [] }],
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string", "minLength": 1, "maxLength": 512 }
          }
        ],
        "responses": {
          "200": {
            "description": "Bounded owner-and-node-scoped binding page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./agent-service-v1.openapi.json#/components/schemas/DialogBindingPage"
                }
              }
            }
          },
          "400": { "description": "Invalid node, query or cursor" },
          "401": { "description": "Session missing or expired" },
          "404": { "description": "agent-service is not configured" },
          "503": { "description": "agent-service or its database is unavailable" }
        }
      }
    },
    "/api/v2/configuration-drafts/validate": {
      "post": { "summary": "Validate configuration draft without network, build or apply effects", "security": [{ "panelSession": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigurationValidateInput" } } } }, "responses": { "200": { "description": "Diagnostics with effectStatus none", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigurationValidationResponse" } } } } } }
    },
    "/api/v2/configuration-drafts/{nodeId}": {
      "get": { "summary": "Reload an owner-scoped raw configuration draft", "security": [{ "panelSession": [] }], "responses": { "200": { "description": "Saved draft", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigurationDraft" } } } }, "404": { "description": "No draft" } } },
      "post": { "summary": "CAS-save a draft; never apply configuration", "security": [{ "panelSession": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigurationSaveInput" } } } }, "responses": { "200": { "description": "Saved draft", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigurationDraft" } } } }, "409": { "description": "Current server version returned without losing local browser text" } } }
    },
    "/api/v2/external-enrollment-hosts": {
      "get": {
        "summary": "List safe host projections available for external Harness enrollment",
        "security": [{ "panelSession": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string", "minLength": 1, "maxLength": 512 }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner-scoped host metadata without credentials or transport refs",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EnrollmentHostPage" } } }
          },
          "400": { "description": "Invalid pagination query" },
          "401": { "description": "Session missing or expired" },
          "404": { "description": "External enrollment is not configured" },
          "503": { "description": "agent-service unavailable" }
        }
      }
    },
    "/api/v2/external-enrollments": {
      "post": {
        "summary": "Enroll one existing Harness without creating or restarting its process",
        "security": [{ "panelSession": [] }],
        "parameters": [
          {
            "name": "X-Panel-CSRF",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "minLength": 43, "maxLength": 43 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExternalEnrollmentRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Exact identity and required capabilities re-observed; registry activation is durable",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExternalEnrollmentResult" } } }
          },
          "400": { "description": "Malformed request" },
          "401": { "description": "Session missing or expired" },
          "403": { "description": "CSRF rejected" },
          "404": { "description": "External enrollment or selected host is not configured" },
          "409": { "description": "Operation, host version, registry, or tunnel binding conflict" },
          "422": { "description": "Endpoint, identity, ownership, schema, readiness, or required capability rejected" },
          "503": { "description": "Outcome unknown; retry the exact operation for reconciliation" }
        }
      }
    },
    "/api/v2/logout": {
      "post": {
        "summary": "Revoke the current Panel session",
        "security": [{ "panelSession": [] }],
        "parameters": [
          {
            "name": "X-Panel-CSRF",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "minLength": 43, "maxLength": 43 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "type": "object", "additionalProperties": false } }
          }
        },
        "responses": {
          "200": { "description": "Session revoked" },
          "401": { "description": "Session missing or expired" },
          "403": { "description": "CSRF rejected" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "panelSession": { "type": "apiKey", "in": "cookie", "name": "__Secure-panel_session" }
    },
    "schemas": {
      "Session": {
        "type": "object",
        "additionalProperties": false,
        "required": ["user", "csrf", "writes_enabled", "inventory_enabled", "enrollment_enabled"],
        "properties": {
          "user": {
            "type": "object",
            "additionalProperties": false,
            "required": ["id", "login", "name"],
            "properties": {
              "id": {
                "type": "string",
                "description": "Server-derived ownerId from the signed Harness registry."
              },
              "login": { "type": "string", "description": "Authenticated reverse-proxy account." },
              "name": { "type": "string" }
            }
          },
          "csrf": { "type": "string", "minLength": 43, "maxLength": 43 },
          "writes_enabled": {
            "type": "boolean",
            "description": "Whether Harness commands are enabled. No YouTrack mutation is exposed."
          },
          "inventory_enabled": {
            "type": "boolean",
            "description": "Whether the private agent-service inventory channel is configured."
          },
          "enrollment_enabled": {
            "type": "boolean",
            "description": "Whether the private external Harness enrollment orchestrator is configured."
          }
        }
      },
      "EnrollmentHostPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schemaId", "items", "nextCursor"],
        "properties": {
          "schemaId": { "const": "external-harness-enrollment-host-page-v1" },
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["hostId", "hostVersion", "displayName", "transport", "availability"],
              "properties": {
                "hostId": { "type": "string", "format": "uuid" },
                "hostVersion": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 },
                "displayName": { "type": "string", "minLength": 1, "maxLength": 200 },
                "transport": { "enum": ["local", "ssh"] },
                "availability": { "type": "string", "minLength": 1, "maxLength": 64 }
              }
            }
          },
          "nextCursor": { "type": ["string", "null"], "minLength": 1, "maxLength": 512 }
        }
      },
      "ExternalEnrollmentRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schemaId", "operationId", "hostId", "expectedHostVersion", "nodeId", "name", "adapter", "endpointUri", "certificateSHA256"],
        "properties": {
          "schemaId": { "const": "external-harness-enrollment-v1" },
          "operationId": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._:@-]{0,127}$" },
          "hostId": { "type": "string", "format": "uuid" },
          "expectedHostVersion": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 },
          "nodeId": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "minLength": 1, "maxLength": 200 },
          "adapter": { "enum": ["cursor", "codex"] },
          "endpointUri": { "type": "string", "format": "uri", "minLength": 1, "maxLength": 512 },
          "certificateSHA256": { "type": "string", "pattern": "^[0-9a-f]{64}$" }
        }
      },
      "ExternalEnrollmentResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["schemaId", "operationId", "nodeId", "status", "registrationRevision", "identityEpoch"],
        "properties": {
          "schemaId": { "const": "external-harness-enrollment-result-v1" },
          "operationId": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._:@-]{0,127}$" },
          "nodeId": { "type": "string", "format": "uuid" },
          "status": { "const": "ready" },
          "registrationRevision": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 },
          "identityEpoch": { "type": "integer", "minimum": 1, "maximum": 9007199254740991 }
        }
      }
    }
  }
}